Commit 7: storefront/storefront: storefront2/commit_message_index.txt: storefront2/store/__pycache__/views.cpython-311.pyc: storefront2/store/views.py

+++ storefront2/store/views.py
@@ -38,6 +38,8 @@ def product_details(request, id):
         return Response(serializer.data)
 
     elif request.method =='DELETE':
+        if product.orderitem_set.count()>0:
+            return Response(status=status.HTTP_405_METHOD_NOT_ALLOWED)
         product.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
